% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro.R
\name{daylight}
\alias{daylight}
\alias{sunsetangle}
\alias{declination}
\alias{rel_dist}
\title{Astronomic helper functions}
\usage{
daylight(lat, day)

sunsetangle(lat, day)

declination(day)

rel_dist(day)
}
\arguments{
\item{lat}{latitude}

\item{day}{julian day (day of the year)}

\item{deg}{angle in degrees}
}
\value{
\code{sunsetangle}: sunset angle in degrees.

\code{declination}: solar declination in degrees.

\code{rel_dist}: relative distance earth-sun.
}
\description{
Some useful calculations of the astronomic configuration of earth in relation to the sun.

sunsetangle
}
\details{
Astronomic helper functions
}
\section{Details}{

  \code{daylight} calculates the duration of the daylight period for a given date and location. It is based purely on geometrical calculation (i.e. does account for dispersion etc.).


  \code{rh2dewpoint} calculates the sunset hour angle for a given latitude and date.


  \code{rh2dewpoint} calculates the solar declination angle in degrees for a given day.


  \code{rel_dist} calculates the relative distance of earth from sun in relation to the mean distance.
}

\examples{
sunsetangle(50,1)
sunsetangle(50,1)
rel_dist(50,1)
}
