% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{magnus}
\alias{magnus}
\title{magnus
calculates the water vapour pressure (over liquid water surfaces) in kPa for a given temprature.}
\usage{
magnus(temp)
}
\arguments{
\item{temp}{temperature in °C}

\item{p}{pressure in kPa}
}
\value{
\code{magnus} .
}
\description{
magnus
calculates the water vapour pressure (over liquid water surfaces) in kPa for a given temprature.
}
\examples{
magnus(100)
}
