% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_estimate.R
\name{etp_estimate}
\alias{etp_estimate}
\title{Estimate potential evapo-transpiration from weather data (1948)}
\usage{
etp_estimate(...)
}
\arguments{
\item{lat}{latitude (used for calculating daylight hours)}

\item{weather}{weather in a data.frame}

\item{columns}{identification of columns in the weather data.frame}

\item{frequency}{output time unit: daily or monthly}
}
\description{
NOT YET IMPLEMENTED!

...

Estimation method can selected manually or will be chosen automatically based on availability of weather data

...
}
\examples{
bla()
}
