% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_hs.R
\name{etp_hs}
\alias{etp_hs}
\title{Potential evapo-transpiration sensu Hargreaves-Samani}
\usage{
etp_hs(lat, day, tmin, tmax, tmean = (tmax + tmin)/2, alt, ke = 0.0135,
  kt = 0.16, kmethod, ...)
}
\arguments{
\item{tmin}{minimum air temperature (°C)}

\item{tmax}{maximum air temperature (°C)}

\item{tmean}{mean air temperature (°C). Deafults to (tmin+tmax)/2}

\item{alt}{elevation above sea level (m)}

\item{ke}{calibration coefficient for Hargreaves' ETP function}

\item{kt}{calibration coefficient for Hargreaves' irradiance estimation.}

\item{kmethod}{Method to use for estimating the kt parameter in the Hargreaves equation. One of "anadale", "allen", "samani".}
}
\value{
Potential evapo-transpiration in mm per period.
}
\description{
Calculates potential evapo-transpiration from air temperature extremes using the approach by Hargreaves&Samani (1985) and popular corrections.
}
\examples{
etp_hs(lat=20,day=45,tmin=1,tmax=7)
}
