% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_pm.R
\name{etp_pm}
\alias{etp_pm}
\title{Potential evapo-transpiration sensu Penman-Monteith}
\usage{
etp_pm(rad, shf = 0, temp = 20, vpd, rh, p = 101.3, r_s = rs(...),
  r_a = ra(...), lh = 2.501 - 0.002361 * temp, ...)
}
\arguments{
\item{rad}{net radiation (MJ/(m2*day))}

\item{shf}{soil heat flux (MJ/(m2*day))}

\item{temp}{air temperature (dC)}

\item{rh}{lative air humidity (\%)}

\item{p}{air pressure (kPa)}

\item{r_s}{canopy surface resistance (s/m)}

\item{r_a}{canopy aerodynamic resistance in s/m}

\item{lh}{latent head of vapourisation}

\item{vp}{air water vapour pressure (kPa)}

\item{lai}{leaf area index}
}
\value{
\code{ra} Potential evapo-transpiration in mm/d
}
\description{
Calculates daily potential evapo-transpiration in mm.
}
\examples{
etp_pm(rad=5,shf=0,temp=4,rh=80)
}
