% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_pt.R
\name{etp_pt}
\alias{etp_pt}
\title{Priestley-Taylor}
\usage{
etp_pt(rad, shf = 0, temp = 20, alpha = 1.26, p = 101.3,
  lh = 2.501 - 0.002361 * temp, ...)
}
\arguments{
\item{rad}{net radiation (MJ/(m²*day)).}

\item{shf}{soil heat flux (MJ/(m²*day)).}

\item{temp}{air temperature (°C).}

\item{p}{air pressure (kPa)}
}
\value{
\code{ra} Potential evapo-transpiration in mm/d.
}
\description{
WARNING: NOT CHECKED.

...

Calculates daily potential evapo-transpiration in using the approach by Priestley-Taylor.
}
\examples{
etp_pt(rad=5,shf=0,temp=20)
}
