% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_th.R
\name{etp_th}
\alias{etp_th}
\title{Potential evapo-transpiration sensu Thorntwaite (1948)}
\usage{
etp_th(lat, dates, temp, fillmethod = na.approx)
}
\arguments{
\item{lat}{latitude (used for calculating daylight hours)}

\item{temp}{air temperature (°C) for the given dates.}

\item{fillmethod}{na.approx or na.spline, used inter- and extrapolate insufficient values to the whole year.}

\item{date}{either a vector of dates (class "Date") or a julian day (numeric). Since the Thorntwaite method is based on an annual heat index, dates should cover the whole year.}
}
\value{
\code{ra} Potential evapo-transpiration in mm/d.
}
\description{
WARNING: IMPLEMENTATION IS PRELIMINARY. RESULTS PROBABLY WRONG. Especially for higher latitudes.

Calculates monthly potential evapo-transpiration in using the approach by Thorntwaite (1948).
}
\examples{
etp_th(lat=54,temp=c(1,14,1),dates=c(1,210,365))
}
