% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermo.R
\name{magnus}
\alias{magnus}
\alias{rh2dewpoint2}
\alias{rh2vp}
\alias{vp2rh}
\alias{dewpoint2rh}
\alias{psychro}
\alias{satpressslope}
\alias{alt_pressure}
\alias{rsurf_}
\alias{ra_}
\alias{wind2m}
\title{Thermodynamic helper functions}
\usage{
magnus(temp)

rh2dewpoint2(temp, rh)

rh2vp(temp, rh)

vp2rh(temp, vp)

dewpoint2rh(temp, dp)

dewpoint2rh(temp, dp)

psychro(p)

satpressslope(temp)

alt_pressure(altitude)

rsurf_(LAIact = 1.44, rstom = 100)

ra_(zom, zoh, d = 2 * h/3, uz = 2.8, h = 0.12, zm = 2, zh = 2)

wind2m(windz, z)
}
\arguments{
\item{temp}{air temperature in °C}

\item{rh}{relative air humidity in \%.}

\item{vp}{water vapour partial pressure in kPa}

\item{dp}{dewpoint temperature in °C}

\item{p}{air pressure in kPa}

\item{LAIact}{active LAI in m²/m².}

\item{rstom}{stomatal resistance in s/m.}

\item{zom:}{roughness length for momentum transfer (m)}

\item{zoh:}{roughness length for heat transfer (m)}

\item{d:}{zero plane displacement height (m)}

\item{uz:}{wind measured at zm (m/s)}

\item{h:}{crop height (m)}

\item{zm:}{height of wind measurement (m)}

\item{zh:}{height of humidity measurements (m)}
}
\value{
\code{rh2dewpoint}: returns dewpoint temperature in °C

\code{rh2vp}: water vapour pressure in kPa.

\code{vp2rh}: relative humidity in \%.

\code{rh2dewpoint} relative air humidity in \%.

\code{dewpoint2vp} water vapur pressure in kPa.

psychrometric constant in kPa/°C

\code{satpressslope}: slope of the curve in kPa/°C

\code{alt_pressure}: air pressure in kPa.

\code{rsurf}: canopy surface resistance in s/m (=stomatal resistance/active LAI). Default values for FAO grass reference surface.

\code{ra}: aerodynamic resistance in s/m. Default arguments for canopy represent the FAO grass reference surface.

\code{wind2m}:
}
\description{
Various useful calculations and conversion functions.
}
\section{Details}{

  \code{magnus} calculates the saturation water vapour pressure (over liquid water surfaces) in kPa for a given temperature using the magnus formula.


  \code{rh2dewpoint} calculates the dewpoint temperature for a given relative air humidity and temperature.


  \code{rh2vp} calculates the water vapour pressure from relative air humidity and temperature.


  \code{vp2rh} calculates the relative air humidity from water vapour pressure and temperature.


  \code{rh2dewpoint} calculates the relative air humidity for a given dewpoint and air and temperature.


  \code{dewpoint2vp} calculates the relative air humidity for a given dewpoint and air and temperature.


  \code{psychro} calculates the psychrometre constant (water in air) for a given air pressure.


  \code{satpressslope} calculates the slope of the water vapour saturation pressure curve in kPa/°C for a given temperature in °C.


  \code{alt_pressure} calculates the atmospheric air pressure at normal conditions for a given altitude above the sea level.


  \code{rsurf}


  \code{rsurf}


  \code{wind2m}
}

\examples{
rh2dewpoint(50,20)
rh2vp(50,20)
rh2vp(50,20)
rh2dewpoint(50,20)
psychro(101.3)
satpressslope(20)
alt_pressure(1000)
rsurf()
ra()
}
