% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra.R
\name{ra}
\alias{ra}
\title{Aerodynamic resistance}
\usage{
ra(wz = 2, zrm = NULL, zrh = NULL, d, h = 0.12, zw = 2, zh = 2)
}
\arguments{
\item{wz}{wind speed at height zw (default 2m/s)}

\item{zrm}{roughness length governing momentum transfer in m (default 0.123*h)}

\item{zrh}{roughness length governing heat transfer in m (default 0.1*zrm)}

\item{d}{zero plane displacement height (default 2/3*h)}

\item{h}{crop height in m (defaults to 0.12m).}

\item{zw}{height of wind measurements in m (default 2m).}

\item{zh}{height of humidity measurements in m (default 2m).}

\item{lai}{Leaf area index}
}
\value{
\code{ra} returns the aerodynamic resistance
}
\description{
Calculation of the aerodynamic resistance for a given crop surface. Defaults to the FAO grass reference surface.
}
