% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_toa}
\alias{rad_toa}
\alias{rad_clearsky}
\alias{rad_nsw}
\alias{rad_nlw}
\alias{ssd2rad}
\title{Radiation balance}
\usage{
rad_toa(lat, day)

rad_clearsky(lat, day, angstr_a = 0.25, angstr_b = 0.5, alt = 0)

rad_nsw(albedo = 0.23, rad = NULL, sunshineduration = NULL, ...)

rad_nlw(tmin, tmax, vp, rad, rad_actual, rad_clearsky, rad_rel)

ssd2rad(lat, day, angstr_a = 0.25, angstr_b = 0.5, sunshinehours = 0)
}
\arguments{
\item{lat}{latitude in degrees}

\item{day}{julian day (day of the year)}

\item{angstr_a}{Ångström a coefficient (defaults to .25)}

\item{angstr_b}{Ångström b coefficient (defaults to .5)}

\item{alt}{altitude above sea level}

\item{vp}{atmospheric water vapour pressure (kPa)}
}
\value{
Daily sum of incoming solar radiation in MJ/m²

\code{rad_clearsky} calculates the daily sum of incoming radiation for a given location under clear sky conditions.

\code{rad_clearsky} daily net sum of incoming shortwave radiation (MJ/m²).

\code{rad_lw_em} daily sum of emitted longwave radition (MJ/m²).

\code{ssd2rad} estimates daily sum of incoming radiation (bottom of atmosphere) from sum of sunshine duration.
}
\description{
Functions to calculate the radiation balance for a given location.
}
\section{Details}{

  \code{rad_toa} calculates the daily sum of incoming extraterrestrial (top of atmosphere) radiation.


  \code{rad_clearsky} calculates the daily sum of incoming radiation (bottom of atmosphere) for a given location under clear sky conditions.


  \code{rad_nsw} #' calculates the daily net amount of shortwvae radiation. This is simply (1-albedo)*irradiance. If no irradiance measurements are available, it can alternatively be estimated from sunshine duration and Ångström coefficients (calling \code{ssd2rad} internally). Future versions will allow estimations based on daily cloud cover values.


  \code{rad_lw_em} estimates daily sum of emitted longwave radiation.


  \code{ssd2rad} estimates daily sum of incoming radiation (bottom of atmosphere) from sum of sunshine duration.
}

\examples{
rad_boa(50,20)
rad_sw()
ssd2rad(50,20)
ssd2rad(50,20)
}
