% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp2rad.R
\name{temp2rad}
\alias{temp2rad}
\title{Estimation of global radiation based on temperature differences}
\usage{
temp2rad(lat, day, tmin, tmax, kt = 0.162, alt = 0, trmax = 0.75,
  method = c("hargreaves"), kmethod = c("allen"))
}
\arguments{
\item{lat}{latitude in degrees}

\item{day}{julian day (day of the year)}

\item{tmin, tmax}{minimum and maximum temperatures for the period in question.}

\item{kt}{regression parameter for Hargreaves formula. 0.162 (default) for interior, 0.19 for coastal regions.}

\item{alt}{altitude in metres above sea level}

\item{trmax}{maximum atmospheric transmissivity (default .75) used for Bristow method. The default value of .75 is the sum of the default Ångström a (.25) and b (.5) coefficients for the direct and diffuce radiation components.}

\item{method}{Method to use: One of "hargreaves" (default) or "bristow".}

\item{kmethod}{Method to use for estimating the kt parameter in the Hargreaves equation. One of "anadale", "allen", "samani".}
}
\value{
Daily sum of incoming solar radiation in MJ/m²
}
\description{
Estimate solar radiation from temperature extremes. Implements some popular equations and adjustments.
}
\section{Details}{

  \code{temp2rad} estimates daily sum of incoming solar radiation (bottom of atmosphere) from air temperature extremes.
  Available methods are sensu Bristow & Campbell 1984 or sensu Hargreaves & Samani1982.
  For the Hargreaves method, the calibration parameter kt can be adjusted o altitude according to Allen (1995), Samani (2000), Annandale (2002).
  In the Bristow 1984 paper, a different method for calculating extraterrestrial radiation is used. This is also applied
#' temp2rad
}

\references{
Bandyopadhyay, A., Bhadra, A., Raghuwanshi, N. S., & Singh, R. (2008). Estimation of monthly solar radiation from measured air temperature extremes. Agricultural and Forest Meteorology, 148(11), 1707–1718. https://doi.org/10.1016/j.agrformet.2008.06.002
Bristow, K. L., & Campbell, G. S. (1984). On the relationship between incoming solar radiation and daily maximum and minimum temperature. Agricultural and Forest Meteorology, 31(2), 159–166. https://doi.org/10.1016/0168-1923(84)90017-0
H. Hargreaves, G., & Samani, Z. (1985). Reference Crop Evapotranspiration From Temperature. Applied Engineering in Agriculture, 1. https://doi.org/10.13031/2013.26773
Maluta, E. N., Mulaudzi, T. S., & Sankaran, V. (2014). Estimation of the Global Solar Radiation on the Horizontal Surface from Temperature Data for the Vhembe District in the Limpopo Province of South Africa. International Journal of Green Energy, 11(5), 454–464. https://doi.org/10.1080/15435075.2013.772518
Ollila, A. (o. J.). Dynamics between Clear, Cloudy and All-Sky Conditions: Cloud Forcing Effects. 20.
Samani, Z. (2000). Estimating Solar Radiation and Evapotranspiration Using Minimum Climatological Data (Hargreaves-Samani equation). 13.
}
